feat(Client): refactoring RecordExtraActions pass only the necessary data instead of all the dataset (#2159)
# Description
ref : #2142 
Instead of passing all the dataset, only the necessary data is passed
through props into the RecordExtraActions component

Please include a summary of the changes and the related issue. Please
also include relevant motivation and context. List any dependencies that
are required for this change.

Closes #(issue_number)

**Type of change**

Please delete options that are not relevant.

- [x] Breaking change (fix or feature that would cause existing
functionality to not work as expected)

**How Has This Been Tested**

Please describe the tests that you ran to verify your changes. And
ideally reference `tests`.

- [x] multilabel
- [x] singlelabel

**Checklist**

- [x] I have merged the original branch into my forked branch
- [ ] I added relevant documentation
- [x] follows the style guidelines of this project
- [x] I did a self-review of my code
- [ ] I added comments to my code
- [x] I made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works

Co-authored-by: keithCuniah <keithcuniah@gmail.comh>