Feature/refactoring 2143 2145 2147 2159 and parents components (#2182)
# Description
this pull request is to solved conflict after merging from Similarity
search feature and to update parents components from theses components

The previous PR are:
- [x] #2143 
- [x] #2145 
- [x] #2147 
- [x] #2159


Please include a summary of the changes and the related issue. Please
also include relevant motivation and context. List any dependencies that
are required for this change.

refs #2142  
**This PR is link to the record parts not the header refactoring**


**Type of change**

Please delete options that are not relevant.
- [x] Breaking change (fix or feature that would cause existing
functionality to not work as expected)

**How Has This Been Tested**

Please describe the tests that you ran to verify your changes. And
ideally reference `tests`.

- [x] multi label
- [x] single label
- [x] TextClassification
- [x] TokenClassification
- [x] Text2Text

**Checklist**

- [x] I have merged the original branch into my forked branch
- [ ] I added relevant documentation
- [x] follows the style guidelines of this project
- [x] I did a self-review of my code
- [x] I added comments to my code
- [ ] I made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works

Co-authored-by: keithCuniah <keithcuniah@gmail.comh>