ci: Remove Pyre from CI (#2358)
Hello!

## Pull Request overview
* Remove `pyre` from the CI.
* Remove the `pyre` configuration file.

## Details
Pyre has not been shown useful. It only causes CI failures, and I don't
think anyone actually looks at the logs. Let's save ourselves to red
crosses and remove it. After all, the CI doesn't even run on the
`develop` branch, only on `main` (and `master` and `releases`).

See
[here](https://github.com/argilla-io/argilla/actions/runs/4184361432/jobs/7249906904)
for an example CI output of the `pyre` workflow.

**Type of change**

- [x] Removal of unused code

**Checklist**

- [x] I have merged the original branch into my forked branch
- [ ] I added relevant documentation
- [x] follows the style guidelines of this project
- [x] I did a self-review of my code
- [ ] I added comments to my code
- [ ] I made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works

---

I intend to create a handful of other PRs to improve the CI situation.
In particular, I want green checkmarks to once again mean that a PR is
likely ready, and for red crosses to mean that it most certainly is not.

- Tom Aarsen