Refactor/remove no workspace usage and better superuser computation (#2373)
# Description

For old argilla versions, workspaces were not required to create
datasets. Some code changes were included to support this behavior

This PR clean these code sections and refactor them for those code parts
where old logic was used. For example, the `user.is_superuser` flag is
computed from the workspace list setup but also persisted to improve the
code usability.

We need still to provide a way to migrate those datasets that have no
dataset to a default workspace. @davidberenstein1957 can you provide a
section in docs where to include these steps from the release notes?

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [x] Refactor (change restructuring the codebase without changing
functionality)
- [ ] Improvement (change adding some improvement to an existing
functionality)
- [ ] Documentation update


**Checklist**

- [x] I have merged the original branch into my forked branch
- [ ] I added relevant documentation
- [x] follows the style guidelines of this project
- [x] I did a self-review of my code
- [x] I added comments to my code
- [ ] I made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works