Docs: Tutorial on image classification (#2420)
# Description

This change introduces a tutorial to work with images in Argilla. 

The tutorials uses these steps:

- Zero shot image classification
- Zero shot text classification
- Bulk labelling with semantic search of images
- Bulk labelling with semantic search of text
- Few shot classification of labelled data using SetFit

The purpose of the tutorial is to show that a labelled dataset combines
the information of both modalities, which facilitates a lighter and more
accurate model.

The tutorial requires a custom build of Argilla that is shared via a
fork, docker image, and HF space. All are described in the tutorial.

This tutorial was discussed with @davidberenstein1957 and @dvsrepo .

**Type of change**

- [x] Documentation update

**How Has This Been Tested**

Tested by running on python 3.8 (colab), 3.9, and 3.10. No automated
testing.

**Checklist**

- [x] I have merged the original branch into my forked branch
- [x] I added relevant documentation
- [x] follows the style guidelines of this project
- [x] I did a self-review of my code
- [x] I made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works