Add datasets database table and Dataset SQLAlchemy model (#2617)
# Description

This PR include early changes to add the new `datasets` database table
and the `Dataset` SQLAlchemy model. I have additionally added also a
factory for dataset so we can create tests later.

There are some points to review like:

* Are dataset names unique?.
* Are dataset guidelines optional?.
* Dataset `status` column could be a boolean column with the name
`draft` in the case that we are sure that we only want to have
draft/not-draft datasets (of course `status` is more flexible if we want
to have different dataset statuses in the future).