Add new API v1 datasets list endpoint (#2624)
# Description

This PR adds the first v1 endpoint for list datasets.

Some changes that we need to discuss:
- I have added `DatasetPolicyV1` class into `policies.py`. We can think
into alternatives to split the policies into different versions if
required.
- I have not used `response_model_exclude_none=True` into the new
endpoint. I think it's not a good idea to remove attributes becase they
are empty and I have taken the chance of not using it with this new v1.
- I have changed how database sessions are managed with the factories.
Now it's possible to work with relationships so for example now the
workspace is automatically created when a dataset factory is used.