Add datasets annotations get enpdoint (#2653)
# Description

This PR adds a new v1 endpoint to get all the annotations belonging to a
specific dataset.

Notes:
* I'm not returning the `dataset_id` on the `Annotation` Pydantic
schema. The reason behind this is because we are requesting annotations
for a dataset so the `dataset_id` is implicit.
* I have created the `Annotation` Pydantic schema inside
`src/argilla/server/schemas/v1/datasets.py`. The reason behind this is
because this is the schema for dataset annotations, if required we can
have another set of schemas for annotations not related with the dataset
(.i.e. we have an endpoint only for annotations).