enhancement: Add `stacklevel` to warning in metrics method (#2635)
Hello!

## Pull Request overview
* Add `stacklevel=2` to `warnings.warn` in all metrics their visualize
methods to give users a better understanding of the source of the
warning.

## Details
### Current situation
Any of the metrics from the [Metrics
docs](https://docs.argilla.io/en/latest/reference/python/python_metrics.html)
can be used for this, but I'll use F1 to show the current behaviour.
Note that `plotly` must not be installed.
```python
import argilla as rg
from argilla.metrics.token_classification import f1

rg.init(api_url="https://argilla-span-marker.hf.space/", api_key="team.apikey")

summary = f1(name="span-marker-bert-base-fewnerd-fine-super")
summary.visualize() # will plot three bar charts with the results
```
Results in:
```
[sic]\models.py:32: UserWarning: Please, install plotly in order to use this feature
%>pip install plotly
  warnings.warn("Please, install plotly in order to use this feature\n" "%>pip install plotly")
```
The proper behaviour is for the bottom line to *not* point to the
`warnings.warn`. This can be modified using the `stacklevel` argument,
which is 1 by default. For this PR, I've incremented it to 2, making it
point one position higher in the call stack.

## After this PR
Since these changes, the output for that same script is:
```
[sic]\argilla_metrics.py:7: UserWarning: Please, install plotly in order to use this feature:
 > pip install plotly
  summary.visualize() # will plot three bar charts with the results
```

---

**Type of change**

- [x] Improvement (change adding some improvement to an existing
functionality)

**How Has This Been Tested**
Running the above script, no test has been added for this.

**Checklist**

- [x] I have merged the original branch into my forked branch
- [ ] I added relevant documentation
- [x] follows the style guidelines of this project
- [x] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)

It is my belief that this change is not notable enough for a CHANGELOG
entry. Let me know if you feel otherwise.

- Tom Aarsen