Add datasets annotations create endpoint (#2657)
# Description

This PR adds a new v1 endpoint to create annotations for a specific
dataset.

Notes:
* It includes the `text` and `rating` annotation types.
* `text` annotation settings is empty right now, we can add possible
settings in the future if required. At the same time if a users
specifies some settings values for this type of annotation they are
discarded.
* `rating` annotation settings allow a `options` attribute that can
store a list of objects. Every object has the structure `{"value": int}`
only allowing integer values.
* `rating` annotation settings allow a `options` attribute with a
minimum of 2 items and a maximum of 100. If it doesn't has many sense we
can change these values.
* I have tried to use Pydantic's [discriminated
unions](https://docs.pydantic.dev/usage/types/#discriminated-unions-aka-tagged-unions)
to select between different models based on the `type` attribute for the
annotations but without any luck. Looks like it has a bug and JSON
serialization is not allowed for subclases using this. I have solved
this using a dynamic validator but please tell me if there is another
better way to solve this.