feat: Allow partial record update (#2576)
# Description

Changes in this PR affect directly the way the clients log data. If the
record info contains the `id`, the record will be partially updated.
This means that logging data can be done by passing a subset of
per-record attributes, instead of the whole record.

Closes #2535

Refs #2534

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [x] Improvement (change adding some improvement to an existing
functionality)

**How Has This Been Tested**

Added tests covering this new feature

**Checklist**

- [x] I have merged the original branch into my forked branch
- [ ] I added relevant documentation
- [x] follows the style guidelines of this project
- [x] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)