feat: Allow exclude vectors when loading records (#2668)
# NOTE

~~PR #2576 must be merged first to make this PR "mergeable"~~

# Description

Work implemented in https://github.com/argilla-io/argilla/pull/2577 and
https://github.com/argilla-io/argilla/pull/2540 tries to introduce a
mechanism for:

1. Loading data with field projections (selecting which record fields
will be loaded)
2. Allow partial updates when uploading projected records

Some problems of this flow are raised during the work:

- There is a field mismatch between client and API records. The proposed
field projection is referring to the Record API data model, and aligning
it with client representation can be messy and give an unexpected
behavior
- The update flag mechanism is a bit cumbersome because of the client
data projection and the record class validation alignments. So, even for
a single field update, you need to retrieve extra fields to keep aligned
the validations.

The field projection in `rg.load` function can introduce confusion
because both record models, client, and API, are not aligned. So, I
prefer to remove this functionality. Instead of that, we can introduce
some extra parameters to load for setup if users want to exclude
`vectors` or `metrics` when loading data. This will help to mitigate
problems founds in #2398

This option, together with changes introduced in #2576, allows users to
load partial record info and still update it without losing stored data.

For example, if users want to add some metadata to already stored
records, they could try the following flow:

```python
ds = rg.load(name=dataset, exclude_vectors=True)

for record in ds:
   record.metadata = ...

rg.log(name=dataset, records=ds) # This will update records only by adding the metadata info
```

Closes [#2398](https://github.com/argilla-io/argilla/issues/2398)

Refs: https://github.com/argilla-io/argilla/pull/2577 and
https://github.com/argilla-io/argilla/pull/2540

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [x] Improvement (change adding some improvement to an existing
functionality)


**How Has This Been Tested**

(Please describe the tests that you ran to verify your changes. And
ideally, reference `tests`)

Existing tests are been modified including this new behavior

**Checklist**

- [x] I have merged the original branch into my forked branch
- [ ] I added relevant documentation
- [x] follows the style guidelines of this project
- [x] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)

---------

Co-authored-by: Tom Aarsen <37621491+tomaarsen@users.noreply.github.com>