Fix : frontend api endpoints with items (#2740)
# Description
Front part to adapt all apiv1 enpoints to a accept `items` 

Please include a summary of the changes and the related issue. Please
also include relevant motivation and context. List any dependencies that
are required for this change.

Ref #2735 

This PR have to be merge with #2716 

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Refactor (change restructuring the codebase without changing
functionality)
- [ ] Improvement (change adding some improvement to an existing
functionality)
- [ ] Documentation update

**How Has This Been Tested**

(Please describe the tests that you ran to verify your changes. And
ideally, reference `tests`)

- [x] datasets list page
- [x] feedback task page

=>3 vue components are concerned :

page of datasets list (frontend/pages/datasets/index.vue)
page of feedback task
(frontend/pages/dataset/_id/annotation-mode/index.vue)
the main content of the page of feedbacktask
(frontend/components/page-contents/feedback-task-content)

**Checklist**

- [x] I have merged the original branch into my forked branch
- [ ] I added relevant documentation
- [x] follows the style guidelines of this project
- [x] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)

Co-authored-by: keithCuniah <keithcuniah@gmail.comh>