Add dataset fields (#2751)
# Description

This PR adds the following changes:
* Add a new database migration creating `fields` table.
* Add a SQLAlchemy `Field` model.
* Add a first endpoint to list fields for a specific dataset `GET
/api/v1/datasets/{dataset_id}/fields`

Notes:
* I have finally used `fields` instead of `field_settings` because the
naming started to feel strange when creating things like Pydantic
schemas. I'm open to suggestions for changing it.
* I didn't add an `Union` for `settings` on `Field` schema. It's because
right now it only has one type that is `text`. We can add the `Union`
and discriminator once that we have more than one type of field.

Closes #2750