Admin users can create records with responses for several users (#2906)
# Description

WARNING !!! This PR changes the records creation endpoint.

Instead of passing a single response, now the endpoint accepts a list of
user responses. This allows admins to restore a backup of the dataset.

@alvarobartt Once this PR is merged in the feature branch, you should
adapt the add_records method to accept a list of users' responses. See
[here](https://github.com/argilla-io/argilla/blob/6fe3d3c2f9c958435e660d192e081203453bf1cb/src/argilla/server/schemas/v1/datasets.py#L228)
for the schema definition.

Closes #2903