fix: `POST /api/v1/datasets` doesn't handle non-existing workspace ids (#2754)
# Description

I've updated the `POST /api/v1/datasets` endpoint to check if the
`workspace_id` provided in the `body` of the request exists. If there is
no `workspace` object associated to the provided id, the endpoint will
return `404 Not Found`. I've also added a new unit test to cover this
case.

Closes #2753 

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Refactor (change restructuring the codebase without changing
functionality)
- [x] Improvement (change adding some improvement to an existing
functionality)
- [ ] Documentation update

**How Has This Been Tested**

- Manually:
- I tried to create a dataset using the `POST /api/v1/datasets` endpoint
and providing a non-existing `workspace_id`. I got `404`. No new rows
were added to the DB.
- Unit Test

**Checklist**

- [x] I have merged the original branch into my forked branch
- [ ] I added relevant documentation
- [x] follows the style guidelines of this project
- [x] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)

---------

Co-authored-by: Francisco Aranda <francis@argilla.io>