Removing extra token classification metrics in backend (#3045)
# Description

The purpose of this PR is to reduce the number of generated properties
used for token classification metrics computation. Sometimes these extra
properties raise index errors as described in issue #1499

With these changes, the `tokens_length`, `token_lenght`,
`mention_lengh`, and `entity_density ` metrics will return no data from
new servers.

Refs https://github.com/argilla-io/argilla/issues/1499

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [x] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [x] Improvement (change adding some improvement to an existing
functionality)

**How Has This Been Tested**

Current tests have been adapted to the changes

**Checklist**

- [x] I have merged the original branch into my forked branch
- [ ] I added relevant documentation
- [ ] follows the style guidelines of this project
- [x] I did a self-review of my code
- [x] I made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)