refactor: mocking search engine for api tests (#3063)
# Description

This PR uses a mock search_engine for all API tests and keeps the
elastic search implementation tests in the proper test suite.

TODO: Some tests are missed for the elastic search engine, but they will
be included in a separate PR with other enhancements.


**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [x] Refactor (change restructuring the codebase without changing
functionality)

**How Has This Been Tested**

Tests have been updated. 

**Checklist**

- [x] I have merged the original branch into my forked branch
- [x] follows the style guidelines of this project
- [x] I did a self-review of my code
- [x] My changes generate no new warnings

---------

Co-authored-by: gabrielmbmb <gmartinbdev@gmail.com>