Adding search reindex for feedback dataset (#3200)
# Description

This PR describes how to update the search indices for feedback datasets
in order to resolver or update missing search parameters when searching
from UI.


**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [x] Documentation update

**How Has This Been Tested**

(Please describe the tests that you ran to verify your changes. And
ideally, reference `tests`)

Documentation changes are reflected properly.

**Checklist**

- [x] I have merged the original branch into my forked branch
- [x] I added relevant documentation
- [x] follows the style guidelines of this project
- [x] I did a self-review of my code
- [x] I made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)

---------

Co-authored-by: nataliaElv <natalia@argilla.io>