Support for `owner`, `admin` and `annotator` role system (#3104)
# Description

This PR will be the main feature branch for a huge refactor that
includes:

1. Rename the current `admin` role to `owner` keeping all the current
functionality but to this new role name
2. Support for an `admin` role, with operation restrictions. Users with
this role can work on scoped workspaces (must be specifically linked to
those workspaces), and the supported operations will be:
   - Create, update, and delete datasets
   - Copy datasets
   - Create and delete dataset records
- Snapshot and restore feedback datasets (reading and publishing all
responses)

The rest of the operations will be denied for this role (users and
workspace management,...)

Closes #3094 

## TODO

(The rest of the tasks will be tackled as separate PR over this branch) 

- [x] Rename `admin` to `owner` (current default credentials for
quickstart images remain the same)
- [x] Support for `admin` role (review and adapt policies)
- [ ] ~~Update all doc sections with this new flow~~ Tackled in #3212
- [ ] ~~HOWTO guide to migrate the current `admin` role (the simplest
way should be to provide a way to update a user role through a CLI
task)~~ Tackled in #3212


Closes #3094 

**Type of change**

(Please delete options that are not relevant. Remember to title the PR
according to the type of change)

- [x] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [x] Refactor (change restructuring the codebase without changing
functionality)

**How Has This Been Tested**

(Please describe the tests that you ran to verify your changes. And
ideally, reference `tests`)

The tests have been updated

**Checklist**

- [x] I have merged the original branch into my forked branch
- [ ] I added relevant documentation
- [x] follows the style guidelines of this project
- [x] I did a self-review of my code
- [ ] I made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] I have added relevant notes to the CHANGELOG.md file (See
https://keepachangelog.com/)

---------

Co-authored-by: Gabriel Martín Blázquez <gmartinbdev@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>